requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.115.12
greenlet==3.2.3
h11==0.16.0
idna==3.10
lxml==5.4.0
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
requests==2.32.4
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
celery[redis,timezone]==5.3.6
uvicorn==0.29.0
redis==5.0.4
